rename sma module to sma_shm
