+++ modules/wr_smashm/main.sh
@@ -2,10 +2,8 @@
 
 OPENWBBASEDIR=$(cd `dirname $0`/../../ && pwd)
 RAMDISKDIR="${OPENWBBASEDIR}/ramdisk"
-#MODULEDIR=$(cd `dirname $0` && pwd)
 DMOD="PV"
 #DMOD="MAIN"
-Debug=$debug
 
 #For Development only
 #Debug=1
@@ -16,14 +14,11 @@ else
 	MYLOGFILE="${RAMDISKDIR}/nurpv.log"
 fi
 
-
-
-
 openwbDebugLog ${DMOD} 2 "SMA serials: ${smaemdpvid}"
 
-bash "$OPENWBBASEDIR/packages/legacy_run.sh" "modules.sma.device" "inverter" "${smaemdpvid}" "1">>$MYLOGFILE 2>&1
+bash "$OPENWBBASEDIR/packages/legacy_run.sh" "modules.sma_shm.device" "inverter" "${smaemdpvid}" "1">>"$MYLOGFILE" 2>&1
 ret=$?
 openwbDebugLog ${DMOD} 2 "EVU RET: ${ret}"
 
-watt=$(<${RAMDISKDIR}/pvwatt)
-echo ${watt}
+watt=$(<"${RAMDISKDIR}/pvwatt")
+echo "${watt}"
